requirements.txt
Django==2.2
django-filter==2.4.0
djangorestframework==3.12.1
drf-nested-routers==0.92.1
Markdown==3.3.3
pytz==2020.4
sqlparse==0.4.1
